+++ tests/test_rdf.py
@@ -64,7 +64,7 @@ class TestRDF:
         assert cause1.get_index() == 0
         assert cause1.get_index(any_kind=True) == 2
 
-        assert len(list(tree.iter_by_type("effect"))) == 3
+        assert len(list(tree.iterator(kind="effect"))) == 3
 
         # tree.print()
         # print()